+++ test/videodownlinkbandwidthpolicy/NoVideoDownlinkBandwidthPolicy.test.ts
@@ -30,7 +30,6 @@ describe('NoVideoDownlinkBandwidthPolicy', () => {
     it('always returns false', () => {
       expect(policy.wantsResubscribe()).to.be.false;
       policy.updateIndex(emptyVideoStreamIndex);
-      policy.updateCalculatedOptimalReceiveSet();
       expect(policy.wantsResubscribe()).to.be.false;
       const metricReport = new DefaultClientMetricReport(logger);
       metricReport.globalMetricReport.currentMetrics['googAvailableReceiveBandwidth'] = 500;
@@ -43,7 +42,6 @@ describe('NoVideoDownlinkBandwidthPolicy', () => {
     it('returns empty set', () => {
       expect(policy.wantsResubscribe()).to.be.false;
       policy.updateIndex(emptyVideoStreamIndex);
-      policy.updateCalculatedOptimalReceiveSet();
       const idSet = policy.chooseSubscriptions();
       expect(idSet.array()).to.be.deep.equal([]);
     });